requirements.txt
click
xmltodict
requests
dnspython